refactor: reorganize project structure and standardize component naming
- move navigation-related components to src/navigation/
- rename various files for better consistency (DrawerContent, etc.)
- update all relative imports after file relocation
- cleanup unused folder structures
