+++ app/(auth)/index.tsx
@@ -2,10 +2,10 @@ import RegistrationAnimatedView from "@/src/components/_views/RegistrationAnimat
 
 import React from "react";
 import { View } from "react-native";
-import { RegistrationModel } from "../(app)/models/RegistrationModel";
+import { RegistrationViewModel } from "./viewModels/RegistrationViewModel";
 
 export default function Registration() {
-  const { ActiveView } = RegistrationModel();
+  const { ActiveView } = RegistrationViewModel();
   return (
     <View style={{ flex: 1, backgroundColor: "#1E252F" }}>
       <RegistrationAnimatedView />
