+++ src/components/_viewModels/RegistrationAnimatedViewModel.ts
