+++ src/components/_views/RideTypeModal.tsx
@@ -2,14 +2,14 @@ import { calculateRidePrice } from "@/src/utils/pricing";
 import { Ionicons } from "@expo/vector-icons";
 import React from "react";
 import {
-  FlatList,
-  Modal,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
+    FlatList,
+    Modal,
+    StyleSheet,
+    Text,
+    TouchableOpacity,
+    View,
 } from "react-native";
-import { RideTypeModalViewModel } from "../_viewsModels/RideTypeModalViewModel";
+import { RideTypeModalViewModel } from "../_viewModels/RideTypeModalViewModel";
 import RideItem from "./RideItem";
 
 export default function RideTypeModal({
